Migrate module parsing utilities to MeteorPackage class.
